Improve course graph UX with dagre layout and search autocomplete.
Replace the manual node positioning with dagre auto-layout, add database-backed course search with validation, and expand course import coverage so CS and other subjects resolve correctly.

+++ src/app/api/init-db/route.ts
@@ -44,7 +44,7 @@ export async function GET() {
         catalog_course_id TEXT,
         description TEXT,
         academic_level TEXT,
-        credits INTEGER,
+        credits REAL,
         date_start TEXT,
         online_offering BOOLEAN,
         campus_offering BOOLEAN,
